+++ src/constants/analytics.ts
@@ -1,0 +1,25 @@
+import { documentTitles } from "./labels";
+
+export const pageViewData = {
+  home: { name: documentTitles.home, section: "Home" },
+  search: {
+    name: documentTitles.search,
+    section: "Search",
+  },
+  workItem: {
+    name: documentTitles.workItem,
+    section: "Item Details",
+  },
+  editionItem: {
+    name: documentTitles.editionItem,
+    section: "Edition Details",
+  },
+  readItem: {
+    name: documentTitles.readItem,
+    section: "Read Online",
+  },
+  collection: {
+    name: documentTitles.collection,
+    section: "Collection",
+  },
+};