+++ src/types/globals.d.ts
@@ -1,0 +1,3 @@
+interface Window {
+  adobeDataLayer: Array<any> | undefined;
+}